:bug: Fixed the namespace of files that have been moved
